+++ killstats/api/schema.py
@@ -28,3 +28,7 @@ class KillboardIndex(Schema):
 
 class CorporationAdmin(Schema):
     corporation: Optional[dict] = None
+
+
+class AllianceAdmin(Schema):
+    alliance: Optional[dict] = None