+++ killstats/models/__init__.py
@@ -1,3 +1,3 @@
 from .general import EveEntity, General
 from .killboard import Killmail
-from .killstatsaudit import KillstatsAudit
+from .killstatsaudit import AlliancesAudit, CorporationsAudit