+++ killstats/tests/test_api/test_allianceapi.py
@@ -4,11 +4,11 @@ from ninja import NinjaAPI
 
 from django.test import TestCase
 
-from allianceauth.eveonline.models import EveCorporationInfo
+from allianceauth.eveonline.models import EveAllianceInfo
 from app_utils.testing import create_user_from_evecharacter
 
-from killstats.api.killboard import KillboardCorporationApiEndpoints
-from killstats.tests.test_api._killstasts_api import KillstatsEmpty, KillstatsMonth
+from killstats.api.killboard import KillboardAllianceApiEndpoints
+from killstats.tests.test_api._alliance_api import KillstatsEmpty, KillstatsMonth
 from killstats.tests.testdata.load_allianceauth import load_allianceauth
 from killstats.tests.testdata.load_killstats import load_killstats_all
 
@@ -32,13 +32,13 @@ class ManageApiJournalCharEndpointsTest(TestCase):
             ],
         )
         cls.api = NinjaAPI()
-        cls.manage_api_endpoints = KillboardCorporationApiEndpoints(api=cls.api)
+        cls.manage_api_endpoints = KillboardAllianceApiEndpoints(api=cls.api)
 
     def test_killstats_killboard_api_no_entry(self):
         self.maxDiff = None
         # given
         self.client.force_login(self.user)
-        url = "/killstats/api/killboard/month/7/year/2020/corporation/0/"
+        url = "/killstats/api/killboard/month/7/year/2020/alliance/0/"
         # when
         response = self.client.get(url)
         # then
@@ -50,7 +50,7 @@ class ManageApiJournalCharEndpointsTest(TestCase):
         self.maxDiff = None
         # given
         self.client.force_login(self.user)
-        url = "/killstats/api/killboard/month/7/year/2024/corporation/0/"
+        url = "/killstats/api/killboard/month/7/year/2024/alliance/0/"
         # when
         response = self.client.get(url)
         # then
@@ -62,7 +62,7 @@ class ManageApiJournalCharEndpointsTest(TestCase):
         self.maxDiff = None
         # given
         self.client.force_login(self.user)
-        url = "/killstats/api/killboard/month/7/year/2024/corporation/2001/"
+        url = "/killstats/api/killboard/month/7/year/2024/alliance/3001/"
         # when
         response = self.client.get(url)
         # then
@@ -70,16 +70,16 @@ class ManageApiJournalCharEndpointsTest(TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.json(), expected_data)
 
-    def test_get_corporation_admin(self):
+    def test_get_alliance_admin(self):
         self.client.force_login(self.user2)
-        url = "/killstats/api/killboard/corporation/admin/"
+        url = "/killstats/api/killboard/alliance/admin/"
         # when
         response = self.client.get(url)
         # then
         excepted_data = [
             {
-                "corporation": {
-                    "2002": {"corporation_id": 2002, "corporation_name": "Eulenclub"}
+                "alliance": {
+                    "3001": {"alliance_id": 3001, "alliance_name": "Voices of War"}
                 }
             }
         ]
@@ -87,11 +87,11 @@ class ManageApiJournalCharEndpointsTest(TestCase):
         self.assertEqual(response.json(), excepted_data)
 
     @patch(
-        "killstats.api.killboard.corporation.killboard.CorporationsAudit.objects.visible_to"
+        "killstats.api.killboard.alliance.killboard.AlliancesAudit.objects.visible_to"
     )
-    def test_get_corporation_admin_no_visible(self, mock_visible_to):
+    def test_get_alliance_admin_no_visible(self, mock_visible_to):
         self.client.force_login(self.user2)
-        url = "/killstats/api/killboard/corporation/admin/"
+        url = "/killstats/api/killboard/alliance/admin/"
 
         mock_visible_to.return_value = None
 
@@ -102,13 +102,13 @@ class ManageApiJournalCharEndpointsTest(TestCase):
         self.assertContains(response, "Permission Denied", status_code=403)
 
     @patch(
-        "killstats.api.killboard.corporation.killboard.CorporationsAudit.objects.visible_to"
+        "killstats.api.killboard.alliance.killboard.AlliancesAudit.objects.visible_to"
     )
-    def test_get_corporation_admin_exception(self, mock_visible_to):
+    def test_get_alliance_admin_exception(self, mock_visible_to):
         self.client.force_login(self.user)
-        url = "/killstats/api/killboard/corporation/admin/"
+        url = "/killstats/api/killboard/alliance/admin/"
 
-        corp = EveCorporationInfo.objects.get(corporation_id=2001)
+        corp = EveAllianceInfo.objects.get(alliance_id=3001)
 
         mock_visible_to.return_value = [corp, "test"]