+++ killstats/tests/test_models/test_killaudit.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
 
-from killstats.models.killstatsaudit import KillstatsAudit
+from killstats.models.killstatsaudit import AlliancesAudit, CorporationsAudit
 from killstats.tests.testdata.load_allianceauth import load_allianceauth
 from killstats.tests.testdata.load_killstats import load_killstats_all
 
@@ -14,5 +14,9 @@ class TestKillstatsAuditModel(TestCase):
         load_killstats_all()
 
     def test_str(self):
-        self.audit = KillstatsAudit.objects.get(corporation__corporation_id=2001)
+        self.audit = CorporationsAudit.objects.get(corporation__corporation_id=2001)
         self.assertEqual(str(self.audit), "Hell RiderZ's Killstats Data")
+
+    def test_alliance_str(self):
+        self.audit = AlliancesAudit.objects.get(alliance__alliance_id=3002)
+        self.assertEqual(str(self.audit), "Eulen Sigma's Killstats Data")